Use Hotwire for amount yielded slider
